kecamatan kelurahan

+++ application/views/humas/kecamatan/lihat.php
@@ -8,6 +8,16 @@
       </ol>
     </div>
     <div class="page-content">
+         <form role="form" method="POST" action="<?php echo base_url();?>Ckecamatan/search/" >
+         <div class="input-group input-group-sm">
+            <input type="text" class="form-control" value="<?php echo set_value('cari'); ?>" name="cari" placeholder="Cari Nama Kecamatan ...">
+            <span class="input-group-btn">
+              <button class="btn btn-info" type="submit"><i class="fa fa-search"></i></button>
+            </span>
+          </div>
+     </form>
+   </br>
+     <?php echo $this->session->flashdata('pesan'); ?>
       <!-- Panel Basic -->
       <div class="panel">
         <header class="panel-heading">
@@ -38,7 +48,7 @@
                     <a href="<?php echo base_url(); ?>Ckecamatan/update/<?php echo $row->kode_kecamatan;?>"><i class="icon wb-wrench" aria-hidden="true"></i></a>
                   </button>
                   <button type="button" class="btn btn-sm btn-icon btn-flat btn-default" data-toggle="tooltip" data-original-title="Hapus">
-                    <a href="<?php echo base_url(); ?>Ckecamatan/hapus/<?php echo $row->kode_kecamatan;?>"><i class="icon wb-close" aria-hidden="true"></i></a>
+                    <a href="<?php echo base_url(); ?>Ckecamatan/hapus/<?php echo $row->kode_kecamatan;?>" onclick="return confirm('Apakah Anda Yakin Ingin Menghapus <?php echo $row->nama_kecamatan;?>')"><i class="icon wb-close" aria-hidden="true"></i></a>
                   </button>
                 </td>
               </tr>
