+++ application/views/humas/kelurahan/lihat.php
@@ -8,6 +8,16 @@
       </ol>
     </div>
     <div class="page-content">
+       <form role="form" method="POST" action="<?php echo base_url();?>Ckelurahan/search/" >
+         <div class="input-group input-group-sm">
+            <input type="text" class="form-control" value="<?php echo set_value('cari'); ?>" name="cari" placeholder="Cari Nama Kelurahan ...">
+            <span class="input-group-btn">
+              <button class="btn btn-info" type="submit"><i class="fa fa-search"></i></button>
+            </span>
+          </div>
+     </form>
+   </br>
+     <?php echo $this->session->flashdata('pesan'); ?>
       <!-- Panel Basic -->
       <div class="panel">
         <header class="panel-heading">
@@ -29,7 +39,7 @@
             </thead>            
             <tbody>
               <?php $no = $this->uri->segment('3') + 1; ?>
-              <?php foreach ($kelurahan->result() as $row){?>
+              <?php foreach ($kelurahan as $row){?>
               <tr>
                 <td><?php echo $no++ ?></td>
                 <td><?php echo $row->kode_kecamatan;?></td>
@@ -42,7 +52,7 @@
                     <a href="<?php echo base_url(); ?>Ckelurahan/update/<?php echo $row->kode_kelurahan;?>"><i class="icon wb-wrench" aria-hidden="true"></i></a>
                   </button>
                   <button type="button" class="btn btn-sm btn-icon btn-flat btn-default" data-toggle="tooltip" data-original-title="Hapus">
-                    <a href="<?php echo base_url(); ?>Ckelurahan/hapus/<?php echo $row->kode_kelurahan;?>"><i class="icon wb-close" aria-hidden="true"></i></a>
+                    <a href="<?php echo base_url(); ?>Ckelurahan/hapus/<?php echo $row->kode_kelurahan;?>" onclick="return confirm('Apakah Anda Yakin Ingin Menghapus <?php echo $row->nama_kelurahan;?>')"><i class="icon wb-close" aria-hidden="true"></i></a>
                   </button>
                 </td>
               </tr>
